==> R/01-format-data.R <==
library(tidyverse)
library(bibliometrix)
library(here)
library(fs)

pubs <- readFiles(dir_ls(here("data-raw", "plain-text"), regexp = "\\.txt"))
pubs <- convert2df(file = pubs, dbsource = "isi", format = "plaintext")

pubs %>%
  as_tibble(rownames = "identifier") %>%
  select(identifier,
         publication_type = PT,
         authors = AU,
         # book_authors = BA,
         # editors = BE,
         # book_group_authors = GP,
         authors_full_name = AF,
         # book_authors_full_name = BF,
         group_authors = CA,
         document_title = TI,
         publication_name = SO,
         # book_series_title = SE,
         # book_series_subtitle = BS,
         language = LA,
         document_type = DT,
         conference_title = CT,
         conference_date = CY,
         conference_location = CL,
         conference_sponsors = SP,
         conference_host = HO,
         author_keywords = DE,
         keywords_plus = ID,
         abstract = AB,
         author_address = C1,
         reprint_address = RP,
         email_address = EM,
         researcher_id = RI,
         orcid = OI,
         funding_agency = FU,
         funding_text = FX,
         cited_references = CR,
         cited_reference_count = NR,
         wos_times_cited = TC,
         total_times_cited = Z9,
         usage_count_180 = U1,
         usage_count_2013 = U2,
         publisher = PU,
         publisher_city = PI,
         publisher_address = PA,
         issn = SN,
         eissn = EI,
         # isbn = BN,
         source_abbreviation = J9,
         iso_source_abbreviation = JI,
         publication_date = PD,
         year_published = PY,
         volume = VL,
         issue = IS,
         # part_number = PN,
         # supplement = SU,
         special_issue = SI,
         # meeting_abstract = MA,
         beginning_page = BP,
         ending_page = EP,
         article_number = AR,
         doi = DI,
         # book_doi = D2,
         early_access_date = EA,
         # early_access_year = EY,
         page_count = PG,
         categories = WC,
         research_areas = SC,
         delivery_number = GA,
         accession_number = UT,
         pubmed_id = PM,
         open_access_id = OA,
         # esi_highly_cited = HC,
         # esi_hot_paper = HP
  ) %>%
  select(identifier, authors, document_title, publication_name, author_keywords,
         keywords_plus, cited_references, source_abbreviation,
         iso_source_abbreviation, year_published, categories,
         research_areas) %>%
  write_rds(path = here("data/cleaned-pubs.rds"), compress = "gz")
